move read commands under commands/actions
